Editor: Split doc translations to separate .cpp, it's too big
This would crash GCC 14 on Linux x86_32 (tested on Fedora 41 i686).

Splitting the file is just a stopgap solution, eventually if the translations
get bigger we'll have the same problem.
