+++ editor/translations/doc_translation.h
@@ -1,5 +1,5 @@
 /**************************************************************************/
-/*  editor_translation.h                                                  */
+/*  doc_translation.h                                                     */
 /**************************************************************************/
 /*                         This file is part of:                          */
 /*                             GODOT ENGINE                               */
@@ -31,10 +31,5 @@
 #pragma once
 
 #include "core/string/ustring.h"
-#include "core/templates/vector.h"
 
-Vector<String> get_editor_locales();
-void load_editor_translations(const String &p_locale);
-void load_property_translations(const String &p_locale);
-void load_extractable_translations(const String &p_locale);
-Vector<Vector<String>> get_extractable_message_list();
+void load_doc_translations(const String &p_locale);
