+++ editor/translations/editor_translation.cpp
@@ -34,7 +34,6 @@
 #include "core/io/file_access_memory.h"
 #include "core/io/translation_loader_po.h"
 #include "core/string/translation_server.h"
-#include "editor/translations/doc_translations.gen.h"
 #include "editor/translations/editor_translations.gen.h"
 #include "editor/translations/extractable_translations.gen.h"
 #include "editor/translations/property_translations.gen.h"
@@ -106,34 +105,6 @@ void load_property_translations(const String &p_locale) {
 		}
 
 		etl++;
-	}
-}
-
-void load_doc_translations(const String &p_locale) {
-	const Ref<TranslationDomain> domain = TranslationServer::get_singleton()->get_or_add_domain("godot.documentation");
-
-	const DocTranslationList *dtl = _doc_translations;
-	while (dtl->data) {
-		if (dtl->lang == p_locale) {
-			Vector<uint8_t> data;
-			data.resize(dtl->uncomp_size);
-			const int64_t ret = Compression::decompress(data.ptrw(), dtl->uncomp_size, dtl->data, dtl->comp_size, Compression::MODE_DEFLATE);
-			ERR_FAIL_COND_MSG(ret == -1, "Compressed file is corrupt.");
-
-			Ref<FileAccessMemory> fa;
-			fa.instantiate();
-			fa->open_custom(data.ptr(), data.size());
-
-			Ref<Translation> tr = TranslationLoaderPO::load_translation(fa);
-
-			if (tr.is_valid()) {
-				tr->set_locale(dtl->lang);
-				domain->add_translation(tr);
-				break;
-			}
-		}
-
-		dtl++;
 	}
 }
